decouple the Fleet singleton from api implementations.
This allows the lower-level API functions to be invoked _without_ having
to share the *entire* singleton and related dependencies. This allows
for easier reuse/sharing.
