+++ packages/data-sdk/src/api/getFleet.ts
@@ -1,0 +1,17 @@
+import { IFleet } from "../model/IFleet";
+import { Authentication } from "../Authentication";
+import { FORMANT_API_URL } from "../config";
+
+export async function getFleet(id: string): Promise<IFleet> {
+  if (!Authentication.token) {
+    throw new Error("Not authenticated");
+  }
+  const data = await fetch(`${FORMANT_API_URL}/v1/admin/fleets/${id}`, {
+    method: "GET",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: "Bearer " + Authentication.token,
+    },
+  });
+  return await data.json();
+}
